+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Added ExportProcessorType to exporter options
+  ([#1504](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1504))
 * Zipkin tags used for InstrumentationLibrary changed from library.name,
   library.version to otel.library.name, otel.library.version respectively.
 * Sending `service.namespace` as Zipkin tag.